pyproject.toml
[project]
name = "bookstore-api"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Phillipml",email = "contato.phillip.menezes@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.9.0"
factory-boy = "^3.3.3"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
